* Add `linux-arm64` CI builds for OpenCV, FFmpeg, OpenBLAS, FFTW, GSL, Leptonica, Tesseract, and others (issue bytedeco/javacv#1021)

+++ flycapture/cppbuild.sh
@@ -10,7 +10,7 @@ fi
 FLYCAPTURE_VERSION=2.13.3.31
 
 case $PLATFORM in
-    linux-arm*)
+    linux-armhf)
         if [[ ! -f "../../downloads/flycapture.${FLYCAPTURE_VERSION}_armhf.tar.gz" ]]; then
             echo "Please place flycapture.${FLYCAPTURE_VERSION}_armhf.tar.gz in the downloads directory"
             exit 1
@@ -19,6 +19,18 @@ case $PLATFORM in
         tar -xzf ../../downloads/flycapture.${FLYCAPTURE_VERSION}_armhf.tar.gz
         rm -Rf $PLATFORM
         mv flycapture.${FLYCAPTURE_VERSION}_armhf $PLATFORM
+        mv $PLATFORM/lib/C/* $PLATFORM/lib
+        mv $PLATFORM/include/C/* $PLATFORM/include
+        ;;
+    linux-arm64)
+        if [[ ! -f "../../downloads/flycapture.${FLYCAPTURE_VERSION}_arm64.tar.gz" ]]; then
+            echo "Please place flycapture.${FLYCAPTURE_VERSION}_arm64.tar.gz in the downloads directory"
+            exit 1
+        fi
+        echo "Decompressing archives..."
+        tar -xzf ../../downloads/flycapture.${FLYCAPTURE_VERSION}_arm64.tar.gz
+        rm -Rf $PLATFORM
+        mv flycapture.${FLYCAPTURE_VERSION}_arm64 $PLATFORM
         mv $PLATFORM/lib/C/* $PLATFORM/lib
         mv $PLATFORM/include/C/* $PLATFORM/include
         ;;
